+++ src/ltspice_converter/__init__.py
@@ -28,4 +28,4 @@ __all__ = [
     "asc_to_netlist",
 ]
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"